requirements.txt
requests~=2.27.1
jikanpy~=4.3.2
Pillow~=9.0.1